Add binary control file caching to the apt manager
Currently, the control file is extracted from each deb on each freight cache
run and augmented using md5, sha1 and sha256 on the deb. As packages will not
often, if ever change size or content once posted, this seems like a costly
operation without a large benefit. This commit adds a basic caching of those
augmented control files. On the first freight-cache run for a package, the
$PACKAGE.deb-control file will be generated with the sums and filesize of the
package. This file can then be used instead of regenerating the same data on
subsequent runs.